+++ mobile/android/android-components/components/feature/customtabs/src/test/java/mozilla/components/feature/customtabs/CustomTabWindowFeatureTest.kt
@@ -7,8 +7,6 @@ package mozilla.components.feature.customtabs
 import android.app.Activity
 import android.content.ActivityNotFoundException
 import android.graphics.Color
-import android.net.Uri
-import androidx.core.net.toUri
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.browser.state.action.ContentAction
 import mozilla.components.browser.state.state.BrowserState
@@ -17,6 +15,7 @@ import mozilla.components.browser.state.state.CustomTabConfig
 import mozilla.components.browser.state.state.CustomTabMenuItem
 import mozilla.components.browser.state.state.createCustomTab
 import mozilla.components.browser.state.store.BrowserStore
+import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.window.WindowRequest
 import mozilla.components.support.test.any
 import mozilla.components.support.test.ext.joinBlocking
@@ -31,7 +30,6 @@ import org.junit.runner.RunWith
 import org.mockito.Mockito.never
 import org.mockito.Mockito.spy
 import org.mockito.Mockito.verify
-import org.mockito.Mockito.verifyNoInteractions
 
 @RunWith(AndroidJUnit4::class)
 class CustomTabWindowFeatureTest {
@@ -42,17 +40,22 @@ class CustomTabWindowFeatureTest {
     private lateinit var store: BrowserStore
     private val sessionId = "session-uuid"
     private lateinit var activity: Activity
-    private val launchUrlFallback: (Uri) -> Unit = spy { _ -> }
+    private lateinit var engineSession: EngineSession
 
     @Before
     fun setup() {
         activity = mock()
+        engineSession = mock()
 
         store = spy(
             BrowserStore(
                 BrowserState(
                     customTabs = listOf(
-                        createCustomTab(id = sessionId, url = "https://www.mozilla.org"),
+                        createCustomTab(
+                            id = sessionId,
+                            url = "https://www.mozilla.org",
+                            engineSession = engineSession,
+                        ),
                     ),
                 ),
             ),
@@ -63,7 +66,7 @@ class CustomTabWindowFeatureTest {
 
     @Test
     fun `given a request to open window, when the url can be handled, then the activity should start`() {
-        val feature = spy(CustomTabWindowFeature(activity, store, sessionId, launchUrlFallback))
+        val feature = spy(CustomTabWindowFeature(activity, store, sessionId))
         val windowRequest: WindowRequest = mock()
 
         feature.start()
@@ -78,7 +81,7 @@ class CustomTabWindowFeatureTest {
     @Test
     fun `given a request to open window, when the url can't be handled, then handleError should be called`() {
         val exception = ActivityNotFoundException()
-        val feature = spy(CustomTabWindowFeature(activity, store, sessionId, launchUrlFallback))
+        val feature = spy(CustomTabWindowFeature(activity, store, sessionId))
         val windowRequest: WindowRequest = mock()
 
         feature.start()
@@ -86,13 +89,12 @@ class CustomTabWindowFeatureTest {
         whenever(windowRequest.url).thenReturn("blob:https://www.firefox.com")
         whenever(activity.startActivity(any(), any())).thenThrow(exception)
         store.dispatch(ContentAction.UpdateWindowRequestAction(sessionId, windowRequest)).joinBlocking()
-
-        verify(launchUrlFallback).invoke("blob:https://www.firefox.com".toUri())
+        verify(engineSession).loadUrl("blob:https://www.firefox.com")
     }
 
     @Test
     fun `creates intent based on default custom tab config`() {
-        val feature = CustomTabWindowFeature(activity, store, sessionId, launchUrlFallback)
+        val feature = CustomTabWindowFeature(activity, store, sessionId)
         val config = CustomTabConfig()
         val intent = feature.configToIntent(config)
 
@@ -103,7 +105,7 @@ class CustomTabWindowFeatureTest {
 
     @Test
     fun `creates intent based on custom tab config`() {
-        val feature = CustomTabWindowFeature(activity, store, sessionId, launchUrlFallback)
+        val feature = CustomTabWindowFeature(activity, store, sessionId)
         val config = CustomTabConfig(
             toolbarColor = Color.RED,
             navigationBarColor = Color.BLUE,
@@ -120,7 +122,7 @@ class CustomTabWindowFeatureTest {
 
     @Test
     fun `creates intent with same menu items`() {
-        val feature = CustomTabWindowFeature(activity, store, sessionId, launchUrlFallback)
+        val feature = CustomTabWindowFeature(activity, store, sessionId)
         val config = CustomTabConfig(
             actionButtonConfig = CustomTabActionButtonConfig(
                 description = "button",
@@ -142,7 +144,7 @@ class CustomTabWindowFeatureTest {
 
     @Test
     fun `handles no requests when stopped`() {
-        val feature = CustomTabWindowFeature(activity, store, sessionId, launchUrlFallback)
+        val feature = CustomTabWindowFeature(activity, store, sessionId)
         feature.start()
         feature.stop()
 
@@ -151,7 +153,6 @@ class CustomTabWindowFeatureTest {
         whenever(windowRequest.url).thenReturn("https://www.firefox.com")
         store.dispatch(ContentAction.UpdateWindowRequestAction(sessionId, windowRequest)).joinBlocking()
         verify(activity, never()).startActivity(any(), any())
-        verifyNoInteractions(launchUrlFallback)
         verify(store, never()).dispatch(ContentAction.ConsumeWindowRequestAction(sessionId))
     }
 }