+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/ExternalAppBrowserFragment.kt
@@ -71,9 +71,7 @@ class ExternalAppBrowserFragment : BaseBrowserFragment(), UserInteractionHandler
             requireActivity(),
             components.store,
             sessionId!!,
-        ) {
-            // No-op. Client may override this
-        }
+        )
         lifecycle.addObserver(windowFeature)
 
         if (manifest != null) {